requirements.txt
Flask==1.1.2
opencv-python==3.4.7.28
numpy==1.19.1   
flask_cors==3.0.9
easydict==1.9.0
pydantic
python-multipart==0.0.5
aiofiles==0.5.0
gdown
lmdb
pillow
pyyaml
opencv-contrib-python
